Added missing snippets, latex fine-tuning, makefile-setup for snippets
The latex fine tuning is done via a sed-script that processes the generated snippet file. This is a hack, and can be achieved in a cleaner way on isabelle level, by defining output syntax, see Library/LaTeXSugar for examples.

The makefile intentionally does not contain a dependence of the snippets to the Isabelle-theories, to keep it decoupled. To force re-buld of snippets do a
  make clean_snippets
first.